+++ src/config/index.js
@@ -57,7 +57,7 @@ config.blockchainRoute = "/btc";
 config.blockchainCallback = config.domain+config.blockchainRoute;
 config.blockchainConfirmationLimit = 6;
 config.blockchainGapLimit = 20;
-config.blockchain_check_gap = true;
+config.blockchainCheckGap = true;
 // PayPal
 config.paypal_ips = ["64.4.248", "64.4.249", "66.211.168", "66.211.168", "173.0.84", "173.0.84", "173.0.88", "173.0.88", "173.0.92", "173.0.93", "173.0.82", "173.0.81"];
 
@@ -112,7 +112,9 @@ function deploy(environment) {
 
 	config.Twitter = false;
 	config.Twitter_tweeting = false;
-	config.Twitter_tweeting_on_live = false;
+	config.Twitter_deleting_tweet = true;
+	config.Twitter_deleting_tweet_on_live = true;
+	config.Twitter_tweet_on_publish = false;
 	config.remoteDatabase = false;
 
 	config.debugging_blockchain = false;
@@ -143,8 +145,9 @@ function deploy(environment) {
 		// config.debugging_reset_logs = true;
 		// config.debugging_blockchain = true;
 		config.local = true;
-		config.remoteDatabase = true;
+		// config.remoteDatabase = true;
 		config.archive_on_publish = true;
+		// config.backupToOnlyFans = true;
 		// config.debugging_crons = true;
 		// config.debugging_paypal = true;
 		// config.debugging_paypal_reset_plans = true;
@@ -162,7 +165,6 @@ function deploy(environment) {
 		config.PayPal = true;
 		config.PayPal_environment = 'sandbox';
 		config.Twitter = true;
-		config.Twitter_tweeting_on_live = true;
 		config.remoteDatabase = true;
 		// config.archive_on_publish = true;
 		config.backupToOnlyFans = true;
@@ -178,7 +180,6 @@ function deploy(environment) {
 		config.PayPal_syncing_webhooks = false;
 		config.Twitter = true;
 		config.Twitter_tweeting = true;
-		config.Twitter_tweeting_on_live = true;
 		config.backupToOnlyFans = true;
 		config.archive_on_publish = true;
 		// config.remoteDatabase = true;