+++ src/modules/twitter.js
@@ -3,16 +3,13 @@ var config = require('../config/index'),
     moment = require('moment'),
     Twit = require('twit');
 
-var T;
-
 /*
 	connects to Twitter's processes
 */
 var connect = function(callback) {
-	var self = this;
 	logger.log('Connecting to Twitter...');
 	if (!config.Twitter) return callback('Twitter is Disabled');
-	T = new Twit(config.TwitterConfig);
+	var T = new Twit(config.TwitterConfig);
 	T.get('account/verify_credentials',{}, function (err, user) { 
         if (err) return callback(err);
         logger.log('Connected to Twitter: %s', user.name);
@@ -25,16 +22,67 @@ var connect = function(callback) {
 }
 module.exports.connect = connect;
 
-var tweetLive = function(tw) {
-    var self = this;
-	logger.log('Tweeting Live');
-	var tweet = "I\'ve just gone live! "+moment(new Date()).format('D/MM @ H:MM')+" "+config.Twitter_link;
-    logger.debug('tweet: %s', tweet);
-	if (!config.Twitter) return logger.debug('Twitter Disabled');
-	if (!config.Twitter_tweeting) return logger.debug('Not Tweeting');
-	T.post('statuses/update', { 'status': tweet }, function(err) { 
-    	if (err) return logger.warn(err);
+var tweet = function(tw, callback) {
+	logger.log('Tweeting');
+    if (!tw) tw = "I\'ve just gone live! "+moment(new Date()).format('D/MM @ H:MM')+" "+config.Twitter_link;
+    else tw += " "+config.Twitter_link
+    logger.debug('tweet: %s', tw);
+	if (!config.Twitter) return callback('Twitter Disabled');
+	if (!config.Twitter_tweeting) return callback('Not Tweeting');
+	var T = new Twit(config.TwitterConfig);
+    T.post('statuses/update', { 'status': tw }, function(err) { 
+    	if (err) return callback(err);
         logger.log('Live Tweet posted');
+        callback(null);
     });
 }
-module.exports.tweetLive = tweetLive;
+module.exports.tweet = tweet;
+
+var tweetOnPublishLock = false;
+var tweetOnPublish = function(video, callback) {
+    if (tweetOnPublishLock) return callback('Recently Tweeted');
+    logger.log('Tweeting Published Video: ', video.title);
+    var tw = "I\'ve just uploaded another video! "+video.link;
+    logger.debug('tweet: %s', tw);
+    if (!config.Twitter) return callback('Twitter Disabled');
+    if (!config.Twitter_tweeting) return callback('Not Tweeting');
+    var T = new Twit(config.TwitterConfig);
+    T.post('statuses/update', { 'status': tw }, function(err) { 
+        if (err) return callback(err);
+        logger.log('Published Tweet posted');
+        callback(null);
+        tweetOnPublishLock = true;
+        setTimeout(function publishTimeout() {
+            tweetOnPublishLock = false;
+            logger.debug('tweetOnPublishLock released');
+        }, 10000);
+    });
+}
+module.exports.tweetOnPublish = tweetOnPublish;
+
+
+var deleteLiveTweet = function(callback) {
+    logger.log('Deleting Live Tweet');
+    if (!config.Twitter) return logger.debug('Twitter Disabled');
+    if (!config.Twitter_deleting_tweet) return logger.debug('Not Deleting Tweet');
+    // T.get('statuses/user_timeline', { 'user_id': Twitter_user.user_id, 'screen_name': Twitter_user.screen_name, 'count': 20, 'since_id': lastTweet }, function (err, tweets) {
+    var T = new Twit(config.TwitterConfig);
+    T.get('statuses/user_timeline', { 'count': 20 }, function (err, tweets) {
+        if (err) return callback(err);
+        if (tweets.length<=0) return cTallback('Error: No Tweets Found');
+        for (var i=0;i<tweets.length;i++) {
+            // logger.log('status: %s', JSON.stringify(tweets[i], null, 4));
+            if (~tweets[i].text.indexOf("/t.co/"))
+                return T.post('statuses/destroy/:id', { 'id': tweets[i].id_str }, function (err, data, response) {
+                    if (err) return callback(err);
+                    logger.log('Live Tweet destroyed: %s', tweets[i].id_str);
+                    callback(null);
+                });       
+        }
+        callback('Error: No Live Tweet Found');
+    });
+}
+module.exports.deleteLiveTweet = deleteLiveTweet;
+
+
+