+++ changelog.md
@@ -51,12 +51,28 @@
  - rtmp key debugging
  **1.0.3 : 1/14/2019**
  - email debugging
-
-
-
+ **1.0.4 : 2/1/2019**
+ - GoProStream-master/stream.js
+ - Tweeting rearranged to stream.js
+ **1.0.5 : 2/12/2019**
+ - menu.js
+ **2/15/2019**
+ - menu additions: connect
+ - twitter: tweetOnPublish
+ **2/20/2019**
+ - blockchain: generateAddress rework
+ 
  **??/??/??**
  - next live @ & schedule
 
 
 **2.0.0 : Android App : ??/??/??**
  - android rtmp relay?
+
+
+
+
+ to:do
+
+blockchain address recycling
+view: videos -> paginate months